Refactor sdr
